chore: 쿠폰 타입 가드 로직 도메인으로 분리
