+++ src/components/OrderSummary/CouponModal/CouponItem/domain/couponDisabled.domain.ts
@@ -1,11 +1,11 @@
-import { CouponType, DiscountType } from "../../../../types/coupon";
-import { CartItemType } from "../../../../types/response";
+import { CouponType, DiscountType } from "../../../../../types/coupon";
+import { CartItemType } from "../../../../../types/response";
 import {
   isBuyXGetYCoupon,
   isFixedCoupon,
   isFreeShippingCoupon,
   isPercentageCoupon,
-} from "./utils";
+} from "./couponTypeGuards.domain";
 
 type ValidatorParams = {
   type: DiscountType;
