+++ src/components/OrderSummary/CouponModal/CouponItem/utils.ts
@@ -1,31 +1,3 @@
-import {
-  BuyXGetYCoupon,
-  CouponType,
-  FixedCoupon,
-  FreeShippingCoupon,
-  PercentageCoupon,
-} from "../../../../types/coupon";
-
-export function isFixedCoupon(coupon: CouponType): coupon is FixedCoupon {
-  return coupon.discountType === "fixed";
-}
-
-export function isBuyXGetYCoupon(coupon: CouponType): coupon is BuyXGetYCoupon {
-  return coupon.discountType === "buyXgetY";
-}
-
-export function isFreeShippingCoupon(
-  coupon: CouponType
-): coupon is FreeShippingCoupon {
-  return coupon.discountType === "freeShipping";
-}
-
-export function isPercentageCoupon(
-  coupon: CouponType
-): coupon is PercentageCoupon {
-  return coupon.discountType === "percentage";
-}
-
 export function formatDate(dateString: string): string {
   const date = new Date(dateString);
